requirements.txt
pytesseract
opencv	
imutils
pillow
numpy
matplotlib
tensorflow-gpu
jupyter
flask
flask_restplus
pdf2image
scikit-image
cython
protobuf
libsm6
libxext6
libxrender-dev
protobuf-compiler
python-pil
python-lxml